+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Minor Improvements
 
 * Add edit SVG to Icon
+* Supports Ajax call for error validation
 
 # 0.16.0